Fix bug with initialization when setRoute was used
Navigation allows for your navLinks to be populated async by allowing you to call init when you are ready. Previously we handled the case where the route was set from the URL, but not an internal call to setRoute. The internal call to setRoute would get lost. Now by calling "reRoute" in place of the old activate first link code, the reRoute function will be able to set the route appropriately (and fall back on bypass, which takes the place of the code previously which set the route to the initial nav link)